% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobFlfreq1bstr.R
\name{annualfreqplot}
\alias{annualfreqplot}
\title{Plots the frequency curve and calculates the return periods}
\usage{
annualfreqplot(pathmain = NULL, qtT, Nsim, durt, qobs, Qsim, Pint, nsy,
  incond, plotpos = "Gringorten", writeResults = TRUE, PDFplots = TRUE)
}
\arguments{
\item{pathmain}{path to files, not needed if writeResults=FALSE}

\item{qtT}{threshold quantile for POT (Peak over threshold) events}

\item{Nsim}{number of simulations}

\item{durt}{critical duration}

\item{qobs}{observed streamflow data;  dataframe with columns Date ("\%Y-\%m-\%d \%H:\%M:\%S") and Q, if writeResults=TRUE path to csv file with columns date and Q}

\item{Qsim}{dataframe of simulated streamflow and snowmel data (otput from the function hydrolsim) , not needed if writeResults=TRUE}

\item{Pint}{dataframe of simulated precipitation values (daily values)}

\item{nsy}{dataframe of extracted POT rain events for each season}

\item{incond}{dataframe of initial conditions (soil moisture deficit, Qin and SWE) for each season}

\item{plotpos}{te formula which is used for the pplotting positions, possible choices are only  Gringorten and Weibull}
}
\value{
csv file with return periods
}
\description{
Plots the frequency curve and calculates the return periods
}
\examples{
\dontrun{
 d1=d[2:5]
 m1=annualfreqplot(qtT=0.9,Nsim=1000,durt=24,qobs,Pint=h,incond=gincon,nsy=d1,Qsim=gsim,writeResults=FALSE,PDFplots=FALSE)
}
}
