% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h1}
\alias{h1}
\title{Temperature data for stN 16.193 , hourly timestep}
\format{A data frame with:
\describe{
  \item{date}{date column}
   \item{time}{time column}
  \item{T}{ Temperature values }
}}
\usage{
h1
}
\description{
Temperature data for stN 16.193 , hourly timestep
}
\keyword{datasets}
