% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pqruts1mqsln24100000matrsn1lapply.R
\name{hydrolsim}
\alias{hydrolsim}
\title{Simulates discharge values from a set of input P,T files
and initial conditions}
\usage{
hydrolsim(pathmain = NULL, ncl = 4, seasn, param.station, Nsim, int1, Pt, E,
  durt, Area1, kd, slconst = 1, snpSpt = 0.3, ttsnow = -1, Tmax = 0.5,
  Tmin = 0.5, writeResults = FALSE, PDFplots = FALSE)
}
\arguments{
\item{pathmain}{path to files, not needed if writeResults=FALSE}

\item{ncl}{number of cluster, used for parallel processing}

\item{seasn}{season}

\item{param.station}{station number}

\item{Nsim}{number of simulations}

\item{int1}{dataframe of initial conditions (SWE,soil moisture deficit and initial discharge ) for events (otput from the function initconditions)}

\item{Pt}{a matrix with each column giving the input P sequence for a flood event (otput from the function rainsimulate1) not needed if writeResults=TRUE}

\item{E}{a matrix with each column giving the input T sequence for a flood event (otput from the function rainsimulate1) not needed if writeResults=TRUE}

\item{durt}{critical duration}

\item{Area1}{watershed Area}

\item{kd}{coefficient for melting (usually 2-5)}

\item{slconst}{loss due to infiltration (2 constant loss, 1 all of the precipitation infiltrates until a reservoir is filled up) )}

\item{snpSpt}{percent snow above which the snowmodel is activated}

\item{ttsnow}{temperature threshold for snow, below this temperature the snowmodel is activated}

\item{Tmax}{max T for snow melt}

\item{Tmin}{min T for snowmelt}

\item{modelsnow}{type of snow model (Snow.sim or snow)}
}
\value{
csv file with columns Q and snowmelt
}
\description{
Uses the PQRUT model hydrological model and a degree day snowmodel to simulate discharge values
By default it uses parallel processing using 4 cores
}
\examples{
\dontrun{
g= hydrolsim(seasn="Ft",param.station=c(0.2,0.1,10),Nsim=1000,int1=g$SWEFt,Pt=as.matrix(h[[1]]),E=h[[5]],durt=24,Area1=500,kd=2,modelsnow="Snow.sim",slconst=1,snpSpt=0.3,ttsnow=-1,Tmax=0.5,Tmin=0.5,writeResults=FALSE,PDFplots=FALSE)
}
}
