% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initcondPQRUT.R
\name{initcond}
\alias{initcond}
\title{Extracts initial conditions for POT events}
\usage{
initcond(pathmain = NULL, this.station = NULL, Qobs, Td, sl, POTF, POTW,
  POTSp, POTS, SWE, durt, PDFplots = FALSE, writeResults = TRUE)
}
\arguments{
\item{pathmain}{path to files, not needed if writeResults=FALSE}

\item{this.station}{station number,, not needed if writeResults=FALSE}

\item{Qobs}{discharge data; dataframe with columns date ("\%Y-\%m-\%d") and Q,if writeResults=TRUE path to txt file with columns date and Q}

\item{Td}{temperature data; dataframe with columns date("\%Y-\%m-\%d") and Td,if writeResults=TRUE path to txt file with columns date and Td}

\item{sl}{soil moisture deficit data; dataframe with columns date("\%Y-\%m-\%d") and sl,if writeResults=TRUE path to txt file with columns date and sl}

\item{POTF, POTW, POTSp, POTS}{results from criticalduration, not needed if writeResults=TRUE}

\item{SWE}{snow water equivalent data; dataframe with columns date("\%Y-\%m-\%d") and swe,if writeResults=TRUE path to txt file with columns date and swe}

\item{durt}{critical duration}

\item{PDFplots}{if TRUE pdf file with plots is saved in the same directory}
}
\value{
if writeResults is TRUE the results are saved in the directory pathmain if FALSE a list that  contains data tables of initial conditions for the POT flood events is returned
}
\description{
Initial condions (initial discharge, soil moisture deficit and SWE) for POT flood events
}
\examples{
 \dontrun{
 b=initcond(Qobs=Q,sl=sl,POTF=a$POTF,POTW=a$POTW,POTSp=a$POTSp,POTS=a$POTS,SWE=SWE,Td=Td,durt=a$d,PDFplots=FALSE,writeResults=FALSE)
 }
}
