% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P}
\alias{P}
\alias{P}
\title{Precipitation data for station 16.193 (in Norway) , daily timestep}
\format{A data frame with:
\describe{
  \item{date}{date column}
  \item{Pr}{Discharge values in m3/s}
  ...
}}
\usage{
P

P
}
\description{
Precipitation data for stN 16.193 , daily timestep

Precipitation data for stN 16.193 , daily timestep
}
\keyword{datasets}
