% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p1}
\alias{p1}
\title{Precipitation data for stN 16.193 , hourly timestep}
\format{A data frame with:
\describe{
  \item{date}{date column}
 \item{time}{time column}
  \item{Pr}{Precipitation values in mm}
  ...
}}
\usage{
p1
}
\description{
Precipitation data for stN 16.193 , hourly timestep
}
\keyword{datasets}
