% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Qd}
\alias{Qd}
\title{Discharge values for station 16.193 (in Norway) , hourly timestep}
\format{A data frame with:
\describe{
  \item{Date}{date column, NA is used for missing values}
  \item{Q}{Discharge values in m3/s}
  ...
}}
\usage{
Qd
}
\description{
Discharge values for stN 16.193 , hourly timestep
}
\keyword{datasets}
