% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfreqseasn.R
\name{rainPOT}
\alias{rainPOT}
\title{Fit precipitation model}
\usage{
rainPOT(pathmain = NULL, datarainfall = "P.txt", durt, qFtset = 0.9,
  qWtset = 0.95, qSptset = 0.92, qStset = 0.9, distfunc,
  writeResults = TRUE, PDFplots = TRUE)
}
\arguments{
\item{pathmain}{path to file , not needed if writeResults=FALSE}

\item{datarainfall}{precipitation data; dataframe with columns date ("\%Y-\%m-\%d") and Pr,if writeResults=TRUE path to txt file with columns date and Pr}

\item{durt}{critical duration in hours}

\item{qFtset}{quantile threshold for Autumn}

\item{qWtset}{quantile threshold for Winter}

\item{qSptset}{quantile threshold for Spring}

\item{qStset}{quantile threshold for Summer}

\item{distfunc}{distribution function GP or Exponential}

\item{PDFplots}{if TRUE pdf file with plots is saved in the same directory}
}
\value{
if writeResults is TRUE the results are saved in the directory pathmain if FALSE a list that  contains parameters and dataframes that lists the POT for each season is returned
}
\description{
Fits the GP/Exponential distribution to peak over threshold precipitation data by splitting the data into different seasons
}
\examples{
\dontrun{
d=rainPOT(datarainfall=P,durt=24,qFtset=0.9,qWtset=0.95,qSptset=0.92,qStset=0.9,distfunc="GP",writeResults=FALSE,PDFplots=FALSE)
}
}
