% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrainflFt624intmatr100000lapply.R
\name{simulateP}
\alias{simulateP}
\title{Simulates precipitation and temperature sequence.}
\usage{
simulateP(pathmain = NULL, Nsim, durt, distfunc, Pexp, hyet, TempSeq,
  writeResults = FALSE, PDFplots = FALSE)
}
\arguments{
\item{pathmain}{path to files, not needed if writeResults=FALSE}

\item{Nsim}{number of simulations}

\item{durt}{critical duration}

\item{distfunc}{statistical distribution :GP or Exponential}

\item{Pexp}{parameters of the fitted rainfall distribution not needed if writeResults=TRUE}

\item{hyet}{list of dataframes of  hyetographs for each season, output from  stormp not needed if writeResults=TRUE
@param TempSeq list of dataframes of temperature series for each season, output from  temprsim not needed if writeResults=TRUE}
}
\value{
a matrix with each column giving the P/T sequence for a flood event
}
\description{
First, values with daily time resoultion are simulated using either GP or exponential distribution, then temporal patterns are used to
disaggregate the values
By default it uses parallel processing using 4 cores
}
\examples{
\dontrun{
h=simulateP(Nsim=1000,durt=24,distfunc="GP",hyet=h,TempSeq=h1,Pexp=d$par,writeResults=FALSE,PDFplots=FALSE)
}
}
