% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sl}
\alias{sl}
\title{soil moisture deficit data for station 16.193 (in Norway) , daily timestep}
\format{A data frame with:
\describe{
  \item{date}{date column}
  \item{sl}{soil moisture deficit values in mm}
 The data has been generated, using the DDD rainfall-runoff model
}}
\usage{
sl
}
\description{
soil moisture deficit data for stN 16.193 , daily timestep
}
\keyword{datasets}
