% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyetogs.R
\name{stormp}
\alias{stormp}
\title{Extracts a set of hyetographs for POT events}
\usage{
stormp(pathmain = NULL, prpFt, prpWt, prpSpt, prpSt, p1, durt,
  writeResults = FALSE, PDFplots = TRUE)
}
\arguments{
\item{pathmain}{path to data, not needed if writeResults=FALSE}

\item{prpFt, prpWt, prpSpt, prpSt}{dataframes that contain the POT for each season, not needed if writeResults=TRUE path}

\item{p1}{dataframe of precipitation data with 1 hour timestep with columns Date "\%d/\%m/\%Y", time=\%H:\%M:\%S" and value)}

\item{durt}{critical duration}

\item{prp}{output from rainPOT}
}
\value{
if writeResults is TRUE the results are saved in the directory pathmain if FALSE a list that  contains  dataframes of hyetographs for each season is returned
}
\description{
The function extracts hyetographs for the POT rainfall events
}
\examples{
\dontrun{
  h=stormp(prpFt=d$pF,prpWt=d$PW,prpSpt=d$PSp,prpSt=d$PS,p1,durt=24,writeResults=FALSE,PDFplots=FALSE)
}
}
