% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SWE}
\alias{SWE}
\title{snow water equivalent data for stN 16.193 , daily timestep}
\format{A data frame with:
\describe{
  \item{date}{date column}
  \item{SWE}{SWE values in mm}
 The data has been generated, using the DDD rainfall-runoff model
}}
\usage{
SWE
}
\description{
snow water equivalent data for stN 16.193 , daily timestep
}
\keyword{datasets}
