% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Td}
\alias{Td}
\title{Temperature data for station 16.193 (in Norway) , daily timestep}
\format{A data frame with:
\describe{
  \item{date}{date column}
  \item{Td}{ Temperature values }
}}
\usage{
Td
}
\description{
Temperature data for stN 16.193 , daily timestep
}
\keyword{datasets}
