% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmSim.R
\name{temprsim}
\alias{temprsim}
\title{Extracts temperature sequence for POT events}
\usage{
temprsim(pathmain = NULL, incondFt, incondWt, incondSpt, incondSt, durt, tm,
  writeResults = FALSE, PDFplots = FALSE)
}
\arguments{
\item{pathmain}{path to files , not needed if writeResults=FALSE}

\item{incondFt, incondWt, incondSpt, incondSt}{dataframes that contain the POT for each season, not needed if writeResults=TRUE}

\item{durt}{critical duration in hours}

\item{tm}{dataframe of temperature data with 1 hour timestep with columns Date "\%d/\%m/\%Y", time=\%H:\%M:\%S" and value)}
}
\value{
returns csv files that contain temperature series for POT events
}
\description{
Extracts temperature sequence for POT events
}
\examples{
\dontrun{
h1=temprsim(incondFt=b$ntp1F,incondWt=b$ntp1W,incondSpt=b$ntp1Sp,incondSt=b$ntp1S,durt=24,tm,writeResults=FALSE,PDFplots=FALSE)
}
}
